# sql/pgwire: Avoid at least one allocation for most Datum types when using PG's text format