The first commit avoids an allocation for:
- Each `DDecimal` Datum
- Each `DInterval` Datum
- Each element in a `DTuple`
- Each element in a `DArray`

The second commit avoids an allocation for:
- Each `DDate` Datum
- Each `DTimestamp` Datum
- Each `DTimestampTZ` Datum

<!-- Reviewable:start -->

---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/10531)
<!-- Reviewable:end -->