@tamird 2016-11-08
What do the benchmarks say?

<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> mod the comment about replacing `putbuf`

---

Reviewed 3 of 3 files at r3, 2 of 2 files at r4.
Review status: all files reviewed at latest revision, 4 unresolved discussions, all commit checks successful.

---

*[pkg/sql/pgwire/encoding.go, line 162 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW1K5B8w4VJAaFVk0Zz-r2-162:-KW2X5LhmAg-1OSxF4XO:b-i14qwt) ([raw file](https://github.com/cockroachdb/cockroach/blob/08a7f932294b57b5fa9c80ffa5b7e8bef7438167/pkg/sql/pgwire/encoding.go#L162)):*
<details><summary><i>Previously, nvanbenschoten (Nathan VanBenschoten) wrote…</i></summary>
> I'm all ears if you have a more elegant solution to this, keeping in mind `Datum.Format`'s signature.
</details>
Looks OK to me, but given this is where you're headed, perhaps you could fully replace `putbuf` with this variable-sized buffer?

It is used as an argument to PutUint{16,32,64}; those calls can be replaced with `binary.Write`. Trickier would be replacing the `Append{Int,Float}` calls, though perhaps `fmt.Fprtinf` could be used? The last use is `initMsg` which sets the first byte to `0` before calling `wrapped.Write`, but there is no value in saving a call to `Write` here, so that should be easy to replace, too.

---

*[pkg/sql/pgwire/types.go, line 188 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW2YUFxZmizxJ9DdI7q:-KW2YUFybsUpBQEwXHph:bst9ack) ([raw file](https://github.com/cockroachdb/cockroach/blob/407fd675e8ae95f68283f26ede2fc4da856af0e6/pkg/sql/pgwire/types.go#L188)):*
> ```Go
> 
> 	case *parser.DTuple:
> 		tb := &b.variablePutbuf
> ```

just remove this local binding?

---

*[pkg/sql/pgwire/types.go, line 204 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW2YYsfOC8UK6hEpS3n:-KW2YYsfOC8UK6hEpS3o:b-7ropln) ([raw file](https://github.com/cockroachdb/cockroach/blob/407fd675e8ae95f68283f26ede2fc4da856af0e6/pkg/sql/pgwire/types.go#L204)):*
> ```Go
> 
> 	case *parser.DArray:
> 		tb := &b.variablePutbuf
> ```

ditto

---

*[pkg/sql/pgwire/types.go, line 340 at r4](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW2Z5D3ac1RO6C2IPnu:-KW2Z5D3ac1RO6C2IPnv:bfqrg5c) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4e5431483ad979916690efe34c8f42e97b20c3a/pkg/sql/pgwire/types.go#L340)):*
> ```Go
> 
> // formatTs formats t with an optional offset into a format lib/pq understands,
> // appending to the provided tmp buffer if it is large enough and returning the
> ```

s/if it is large enough/ - it always calls `append` so it will reallocate if needed.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-:-KW2YbmDkVp1QsW2QgdX:bbra9f3)*
<!-- Sent from Reviewable.io -->