@nvanbenschoten 2016-11-14
Unfortunately, none of our current benchmarks stress this.

---

Review status: 1 of 3 files reviewed at latest revision, 4 unresolved discussions.

---

*[pkg/sql/pgwire/encoding.go, line 162 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW1K5B8w4VJAaFVk0Zz-r2-162:-KWYvl9dwDQglIuuiEGE:b-76u6ng) ([raw file](https://github.com/cockroachdb/cockroach/blob/08a7f932294b57b5fa9c80ffa5b7e8bef7438167/pkg/sql/pgwire/encoding.go#L162)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> Looks OK to me, but given this is where you're headed, perhaps you could fully replace `putbuf` with this variable-sized buffer?
> 
> It is used as an argument to PutUint{16,32,64}; those calls can be replaced with `binary.Write`. Trickier would be replacing the `Append{Int,Float}` calls, though perhaps `fmt.Fprtinf` could be used? The last use is `initMsg` which sets the first byte to `0` before calling `wrapped.Write`, but there is no value in saving a call to `Write` here, so that should be easy to replace, too.
</details>
I initially considered that exact approach, but it seems like extra overhead and effort in all cases for no real benefit.

---

*[pkg/sql/pgwire/types.go, line 188 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW2YUFxZmizxJ9DdI7q:-KWYwDPkNS3ltj_NHcYy:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/407fd675e8ae95f68283f26ede2fc4da856af0e6/pkg/sql/pgwire/types.go#L188)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> just remove this local binding?
</details>
Done.

---

*[pkg/sql/pgwire/types.go, line 204 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW2YYsfOC8UK6hEpS3n:-KWYwDzApe0dcvigOXQh:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/407fd675e8ae95f68283f26ede2fc4da856af0e6/pkg/sql/pgwire/types.go#L204)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> ditto
</details>
Done.

---

*[pkg/sql/pgwire/types.go, line 340 at r4](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW2Z5D3ac1RO6C2IPnu:-KWYx5ehM4ko8K1aKdnI:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4e5431483ad979916690efe34c8f42e97b20c3a/pkg/sql/pgwire/types.go#L340)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> s/if it is large enough/ - it always calls `append` so it will reallocate if needed.
</details>
Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-:-KWYtxL5i8v-qxLJ9srt:b-vgoce7)*
<!-- Sent from Reviewable.io -->