@tamird 2016-11-14
Reviewed 1 of 3 files at r3, 2 of 2 files at r5, 2 of 2 files at r6.
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[pkg/sql/pgwire/encoding.go, line 162 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW1K5B8w4VJAaFVk0Zz-r2-162:-KWYzfbGo9ZWqI3Zdpkp:b8gjbvg) ([raw file](https://github.com/cockroachdb/cockroach/blob/08a7f932294b57b5fa9c80ffa5b7e8bef7438167/pkg/sql/pgwire/encoding.go#L162)):*
<details><summary><i>Previously, nvanbenschoten (Nathan VanBenschoten) wrote…</i></summary>
> I initially considered that exact approach, but it seems like extra overhead and effort in all cases for no real benefit.
</details>
What overhead? The benefit is just less random crap in the writeBuffer structure.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531)*
<!-- Sent from Reviewable.io -->