@nvanbenschoten 2016-11-15
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks failed.

---

*[pkg/sql/pgwire/encoding.go, line 162 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW1K5B8w4VJAaFVk0Zz-r2-162:-KWdzoZnnbb5QU-UzbOE:bx2ummf) ([raw file](https://github.com/cockroachdb/cockroach/blob/08a7f932294b57b5fa9c80ffa5b7e8bef7438167/pkg/sql/pgwire/encoding.go#L162)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> What overhead? The benefit is just less random crap in the writeBuffer structure.
</details>
Compare `binary.BigEndian.PutUint16` to `binary.Write` and `strconv.AppendInt` to `fmt.Fprintf`. Both alternatives are much larger functions that do a lot more work and usher parameters through interfaces. This was meant as an optimization, so I don't want to risk losing performance.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531)*
<!-- Sent from Reviewable.io -->