@nvanbenschoten 2016-11-16
Review status: 0 of 3 files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[pkg/sql/pgwire/encoding.go, line 162 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531#-KW1K5B8w4VJAaFVk0Zz-r2-162:-KWejoZuHK55YNLrwQCP:b-f5dg6d) ([raw file](https://github.com/cockroachdb/cockroach/blob/08a7f932294b57b5fa9c80ffa5b7e8bef7438167/pkg/sql/pgwire/encoding.go#L162)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> Please add a comment, then. FWIW, this is why (in my opinion) it's not a good idea to optimize without benchmarks, so I'd really prefer that you write a benchmark here. Remember https://github.com/cockroachdb/cockroach/pull/6661? :trollface: 
</details>
Ok, I've added a first commit full of benchmarks. The results are in the commit messages. Improvements were as expected. The result of attempting to use `fmt.Fprintf` was:

```
name                 old time/op    new time/op    delta
WriteTextInt-4         74.8ns ± 3%   199.9ns ± 2%  +167.32%        (p=0.000 n=10+10)
WriteTextFloat-4        205ns ± 2%     535ns ± 4%  +160.65%        (p=0.000 n=10+10)

name                 old alloc/op   new alloc/op   delta
WriteTextInt-4         0.00B ±NaN%     8.00B ± 0%     +Inf%        (p=0.000 n=10+10)
WriteTextFloat-4       0.00B ±NaN%     8.00B ± 0%     +Inf%        (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
WriteTextInt-4          0.00 ±NaN%      1.00 ± 0%     +Inf%        (p=0.000 n=10+10)
WriteTextFloat-4        0.00 ±NaN%      1.00 ± 0%     +Inf%        (p=0.000 n=10+10)
```

This is a clear showstopper for that approach.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10531)*
<!-- Sent from Reviewable.io -->